Add bootstrap tests.
Fix error when writing to TOML file

The commit fixes an error that occurs when writing to a TOML file.
The error is wrapped with a custom error message.
This ensures that the error is properly handled and provides more
context to the developer.

The commit also includes changes to the import statements in the
bootstrap.go file.

Signed-off-by: SammyOina <sammyoina@gmail.com>