Fix Docker image tag and file path in Makefile and Dockerfile
The Docker image tag and file path in the Makefile and Dockerfile have been fixed to use the correct repository name "Mainflux" instead of "sammyoina". This ensures that the Docker images are tagged and pushed to the correct repository. Additionally, the file path in the Dockerfile has been updated to correctly move the built executable to the desired location.

Signed-off-by: SammyOina <sammyoina@gmail.com>